[BREAKING] Version 0.5.0

+++ test/libsolidity/SolidityOptimizer.cpp
@@ -216,7 +216,7 @@ BOOST_AUTO_TEST_CASE(array_copy)
 					data1[i] = msg.data[i];
 				data2 = data1;
 				l = data2.length;
-				y = uint(data2[x]);
+				y = uint(uint40(data2[x]));
 			}
 		}
 	)";
